+++ tests/morphology.cpp
@@ -557,14 +557,36 @@ BOOST_AUTO_TEST_CASE( swc_first_order_sections )
     const brion::Morphology source( path.string( ));
 
     checkEqualArrays( *source.readSections( stage ), 4,
-                 V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 0 ), V2i( 3, 0 ));
-    // The tree construction algorithm reserves the order of two sections
+                      V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 0 ), V2i( 3, 0 ));
+    // The tree construction algorithm reverses the order of the sections
     // compared to how they appear in the file
     checkEqualArrays( *source.readPoints( stage ), 4, V4f( 0, 0, 0, 20 ),
-                      V4f( 0, 0, 1, 4 ), V4f( 0, 0, 3, 4 ), V4f( 0, 0, 2, 4 ));
+                      V4f( 0, 0, 3, 4 ), V4f( 0, 0, 2, 4 ), V4f( 0, 0, 1, 4 ));
     checkEqualArrays( *source.readSectionTypes(), 4,
-                      SOMA, AXON, APICAL_DENDRITE, DENDRITE );
+                      SOMA, APICAL_DENDRITE, DENDRITE, AXON );
 
+}
+
+BOOST_AUTO_TEST_CASE( swc_first_order_sections_from_arbitrary_points )
+{
+    boost::filesystem::path path( BRION_TESTDATA );
+    const brion::MorphologyRepairStage stage = brion::MORPHOLOGY_REPAIRED;
+    path /= "swc/first_order_sections_ring.swc";
+
+    const brion::Morphology source( path.string( ));
+
+    checkEqualArrays( *source.readSections( stage ), 4,
+                      V2i( 0, -1 ), V2i( 5, 0 ), V2i( 8, 0 ), V2i( 11, 0 ));
+    // The tree construction algorithm reverses the order of the sections
+    // compared to how they appear in the file
+    checkEqualArrays( *source.readPoints( stage ), 14, V4f( 0, 0, 1, 0 ),
+                      V4f( 0, 0, 2, 0 ), V4f( 0, 0, 3, 0 ),  V4f( 0, 0, 4, 0 ),
+                      V4f( 0, 0, 5, 0 ), V4f( 0, 0, 4, 0 ), V4f( 3, 1, 10, 1 ),
+                      V4f( 3, 2, 11, 1 ), V4f( 0, 0, 3, 0 ), V4f( 2, 1, 8, 1 ),
+                      V4f( 2, 2, 9, 1 ), V4f( 0, 0, 2, 0 ), V4f( 1, 1, 6, 1 ),
+                      V4f( 1, 2, 7, 1 ));
+    checkEqualArrays( *source.readSectionTypes(), 4,
+                      SOMA, APICAL_DENDRITE, DENDRITE, AXON );
 }
 
 BOOST_AUTO_TEST_CASE( swc_bifurcation )
@@ -594,11 +616,11 @@ BOOST_AUTO_TEST_CASE( swc_end_points )
     const brion::Morphology source( path.string( ));
 
     checkEqualArrays( *source.readSections( stage ), 6,
-                      V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 1 ), V2i( 4, 1 ),
-                      V2i( 7, 0 ), V2i( 8, 0 ));
+                      V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 0 ), V2i( 3, 0 ),
+                      V2i( 4, 3 ), V2i( 6, 3 ));
 
     checkEqualArrays( *source.readSectionTypes(), 6,
-                      SOMA, AXON, AXON, AXON, UNDEFINED, UNDEFINED );
+                      SOMA, UNDEFINED, UNDEFINED, AXON, AXON, AXON );
 }
 
 BOOST_AUTO_TEST_CASE( swc_fork_points )
@@ -610,11 +632,11 @@ BOOST_AUTO_TEST_CASE( swc_fork_points )
     const brion::Morphology source( path.string( ));
 
     checkEqualArrays( *source.readSections( stage ), 6,
-                      V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 1 ), V2i( 4, 1 ),
-                      V2i( 7, 0 ), V2i( 8, 0 ));
+                      V2i( 0, -1 ), V2i( 1, 0 ), V2i( 2, 0 ), V2i( 3, 0 ),
+                      V2i( 4, 3 ), V2i( 6, 3 ));
 
     checkEqualArrays( *source.readSectionTypes(), 6,
-                      SOMA, AXON, AXON, AXON, UNDEFINED, UNDEFINED );
+                      SOMA, UNDEFINED, UNDEFINED, AXON, AXON, AXON );
 }
 
 BOOST_AUTO_TEST_CASE( swc_neuron )
@@ -624,7 +646,7 @@ BOOST_AUTO_TEST_CASE( swc_neuron )
 
     brion::Morphology neuron( path.string( ));
     BOOST_CHECK_EQUAL( neuron.readPoints( brion::MORPHOLOGY_REPAIRED )->size(),
-                       927 );
+                       933 );
     BOOST_CHECK_EQUAL( neuron.getCellFamily(), brion::FAMILY_NEURON );
     BOOST_CHECK( neuron.readPerimeters()->empty( ));
 }